Refactoring python for readability
